+++ examples/data-sources/iosxr_class_map_qos/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_class_map_qos" "example" {
+  class_map_name = "TEST"
+}