+++ examples/data-sources/iosxr_policy_map_qos/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_policy_map_qos" "example" {
+  policy_map_name = "PM1"
+}