+++ examples/data-sources/iosxr_qos_class_map/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_qos_class_map" "example" {
-  class_map_name = "TEST"
-}