+++ examples/data-sources/iosxr_qos_policy_map/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_qos_policy_map" "example" {
-  policy_map_name = "PM1"
-}