+++ examples/resources/iosxr_policy_map_qos/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxr_policy_map_qos.example "Cisco-IOS-XR-um-policymap-classmap-cfg:/policy-map/type/qos[policy-map-name=PM1]"