+++ examples/resources/iosxr_qos_class_map/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_qos_class_map.example "Cisco-IOS-XR-um-policymap-classmap-cfg:/class-map/type/qos[class-map-name=TEST]"