+++ examples/resources/iosxr_qos_policy_map/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_qos_policy_map.example "Cisco-IOS-XR-um-policymap-classmap-cfg:/policy-map/type/qos[policy-map-name=PM1]"